fix: ListCell image 영역 설정

+++ OpenMarket/OpenMarket/Extension/Int+Extension.swift
@@ -9,11 +9,12 @@ extension Int {
     }
 }
 
-extension Double {
+extension Decimal {
     func format() -> String? {
         guard let formattedString = NumberFormatter.decimal.string(for: self) else {
             return nil
         }
+
         return formattedString
     }
 }
