+++ OpenMarket/OpenMarket/Model/Product.swift
@@ -7,9 +7,9 @@ struct Product: Decodable, Hashable {
     var description: String?
     var thumbnail: String
     var currency: Currency
-    var price: Double
-    var bargainPrice: Double
-    var discountedPrice: Double
+    var price: Decimal
+    var bargainPrice: Decimal
+    var discountedPrice: Decimal
     var stock: Int
     var createdAt: Date
     var issuedAt: Date
